requirements.txt
names~=0.3.0
pytz~=2024.1
loguru~=0.7.2
httpx~=0.27.0
urllib3~=2.2.2
art~=6.2
PyYAML~=6.0.2
pydantic~=2.8.2
inquirer~=3.3.0
colorama~=0.4.6
better_proxy
aiofiles~=24.1.0
rich~=13.9.2
psutil~=7.0.0
aiohttp~=3.11.18
eth-account~=0.13.7
eth-typing~=5.2.1
web3~=7.11.0
base58~=2.1.1